agent/requirements.txt
# Core
numpy
pandas
scipy
scikit-learn
pyyaml
python-dotenv

# Chemistry
rdkit-pypi

# ML / RL (install CUDA torch separately as needed)
torch
transformers
accelerate
ray[rllib]
stable-baselines3

# Tracking (optional)
wandb
